Add events for changing permissions.
