+++ src/routes/Project/routes/Permissions/components/RolesTable/RolesTable.enhancer.js
@@ -1,20 +1,20 @@
-import { get, some, omit } from 'lodash'
 import { compose } from 'redux'
 import { connect } from 'react-redux'
 import { withHandlers, withStateHandlers, withProps } from 'recompose'
 import { withFirestore } from 'react-redux-firebase'
 import { withNotifications } from 'modules/notification'
 import { spinnerWhileLoading } from 'utils/components'
-import { triggerAnalyticsEvent } from 'utils/analytics'
 import { withStyles } from '@material-ui/core/styles'
-import { getOrderedRoles, getProject } from 'selectors'
+import { getOrderedRoles, getProject, getAuthUid } from 'selectors'
 import styles from './RolesTable.styles'
+import * as handlers from './RolesTable.hanlders'
 
 export default compose(
   withNotifications,
   withFirestore,
   // Map redux state to props
   connect((state, props) => ({
+    uid: getAuthUid(state, props),
     project: getProject(state, props),
     orderedRoles: getOrderedRoles(state, props)
   })),
@@ -33,61 +33,7 @@ export default compose(
       })
     }
   ),
-  withHandlers({
-    updateRole: ({ firestore, project, projectId, showSuccess }) => async (
-      roleUpdates,
-      dispatch,
-      formProps
-    ) => {
-      const currentRoles = get(project, 'roles', {})
-      await firestore.update(`projects/${projectId}`, {
-        roles: {
-          ...currentRoles,
-          [formProps.roleKey]: {
-            ...get(currentRoles, formProps.roleKey, {}),
-            permissions: roleUpdates
-          }
-        }
-      })
-      showSuccess('Role updated successfully!')
-      triggerAnalyticsEvent('updateRole', {
-        projectId,
-        roleName: formProps.roleKey
-      })
-    },
-    addRole: props => async newRole => {
-      const { firestore, project, projectId } = props
-      const currentRoles = get(project, `roles`, {})
-      if (some(currentRoles, { name: newRole.name })) {
-        const existsErrMsg = 'Role with that name already exists'
-        props.showError(existsErrMsg)
-        throw new Error(existsErrMsg)
-      }
-      await firestore.update(`projects/${projectId}`, {
-        roles: {
-          ...currentRoles,
-          [newRole.name]: {
-            editPermissions: true
-          }
-        }
-      })
-      props.showSuccess('New Role added successfully!')
-      triggerAnalyticsEvent('addRole', { projectId })
-      props.closeNewRole()
-    },
-    deleteRole: ({
-      firestore,
-      project: { roles },
-      projectId,
-      showSuccess
-    }) => async roleKey => {
-      await firestore.update(`projects/${projectId}`, {
-        roles: omit(roles, [roleKey])
-      })
-      showSuccess('Role deleted successfully!')
-      triggerAnalyticsEvent('deleteRole', { projectId, roleKey })
-    }
-  }),
+  withHandlers(handlers),
   withProps(({ newRoleOpen, auth }) => ({
     addRoleDisabled: newRoleOpen
   })),
